Move access traits specific type support from ArborX_DetailsConcepts.hpp to ArborX_Traits.hpp
