+++ src/details/ArborX_Traits.hpp
@@ -11,6 +11,7 @@
 #ifndef ARBORX_ACCESS_TRAITS_HPP
 #define ARBORX_ACCESS_TRAITS_HPP
 
+#include <ArborX_DetailsConcepts.hpp> // is_complete
 #include <ArborX_Point.hpp>
 
 #include <Kokkos_Core.hpp>
@@ -66,6 +67,78 @@ struct Access<View, Tag,
 };
 
 } // namespace Traits
+} // namespace ArborX
+
+namespace ArborX
+{
+namespace Details
+{
+
+template <typename T, typename TTag, typename = void>
+struct has_access_traits : std::false_type
+{
+};
+
+template <typename T, typename TTag>
+struct has_access_traits<
+    T, TTag, std::enable_if_t<is_complete<Traits::Access<T, TTag>>::value>>
+    : std::true_type
+{
+};
+
+template <typename T, typename = void>
+struct has_memory_space : std::false_type
+{
+};
+
+template <typename T>
+struct has_memory_space<
+    T,
+    std::enable_if_t<Kokkos::is_memory_space<typename T::memory_space>::value>>
+    : std::true_type
+{
+};
+
+template <typename Traits>
+struct result_of_get
+{
+  using type = decltype(Traits::get(
+      std::declval<first_template_parameter_t<Traits> const &>(), 0));
+};
+
+template <typename Traits>
+using decay_result_of_get_t =
+    std::decay_t<typename result_of_get<Traits>::type>;
+
+template <typename Traits, typename = void>
+struct has_get : std::false_type
+{
+};
+
+template <typename Traits>
+struct has_get<
+    Traits,
+    std::void_t<decltype(Traits::get(
+        std::declval<first_template_parameter_t<Traits> const &>(), 0))>>
+    : std::true_type
+{
+};
+
+template <typename Traits, typename = void>
+struct has_size : std::false_type
+{
+};
+
+template <typename Traits>
+struct has_size<
+    Traits,
+    std::enable_if_t<std::is_integral<decltype(Traits::size(
+        std::declval<first_template_parameter_t<Traits> const &>()))>::value>>
+    : std::true_type
+{
+};
+
+} // namespace Details
 } // namespace ArborX
 
 #endif
